KG-160. Split event classes and move to a separate package
- Split feature events for remote writer. Moved classes into a separate package;
- Updated related code.
