+++ agents/agents-core/src/commonMain/kotlin/ai/koog/agents/core/feature/model/events/llmCallEvents.kt
@@ -1,0 +1,50 @@
+package ai.koog.agents.core.feature.model.events
+
+import ai.koog.prompt.dsl.ModerationResult
+import ai.koog.prompt.dsl.Prompt
+import ai.koog.prompt.message.Message
+import kotlinx.serialization.Serializable
+
+/**
+ * Represents an event indicating the start of a call to a Language Learning Model (LLM).
+ *
+ * This event captures the details of the LLM interaction at the point of invocation, including the
+ * input prompt and any tools that will be used during the call. It extends the `DefinedFeatureEvent` class
+ * and serves as a specific type of event in a feature-driven framework.
+ *
+ * @property prompt The input prompt encapsulated as a [Prompt] object. This represents the structured set of
+ *                  messages and configuration parameters sent to the LLM.
+ * @property tools The list of tools, represented by their string identifiers, being used within the scope
+ *                 of the LLM call. These tools may extend or enhance the core functionality of the LLM.
+ * @property eventId A string representing the event type.
+ */
+@Serializable
+public data class BeforeLLMCallEvent(
+    val runId: String,
+    val prompt: Prompt,
+    val model: String,
+    val tools: List<String>,
+    override val eventId: String = BeforeLLMCallEvent::class.simpleName!!,
+) : DefinedFeatureEvent()
+
+/**
+ * Represents an event signaling the completion of an LLM (Large Language Model) call.
+ *
+ * This event encapsulates the responses provided by the LLM during its operation. It serves as a
+ * record of the responses generated by the LLM, marking the end of a particular interaction cycle.
+ * The event is used within the system to capture relevant output data and ensure proper tracking
+ * and logging of LLM-related interactions.
+ *
+ * @property responses A list of responses generated by the LLM, represented as instances of [Message.Response].
+ *                     Each response contains content, metadata, and additional context about the interaction.
+ * @property eventId A string representing the event type.
+ */
+@Serializable
+public data class AfterLLMCallEvent(
+    val runId: String,
+    val prompt: Prompt,
+    val model: String,
+    val responses: List<Message.Response>,
+    val moderationResponse: ModerationResult? = null,
+    override val eventId: String = AfterLLMCallEvent::class.simpleName!!,
+) : DefinedFeatureEvent()
