+++ agents/agents-core/src/commonMain/kotlin/ai/koog/agents/core/feature/model/events/nodeEvents.kt
@@ -1,0 +1,73 @@
+package ai.koog.agents.core.feature.model.events
+
+import ai.koog.agents.core.feature.model.AIAgentError
+import kotlinx.serialization.Serializable
+
+/**
+ * Represents an event triggered when the execution of a specific AI agent node starts.
+ *
+ * This event captures the initial execution context of an AI agent's processing node.
+ * It provides information about the node's name, the input being processed,
+ * and a unique identifier for the event.
+ *
+ * Events of this kind are part of the feature event system and are primarily
+ * used for tracking and monitoring purposes, particularly in scenarios where
+ * understanding the flow of agent execution is essential.
+ *
+ * @property nodeName The name of the node whose execution is starting.
+ * @property input The input data being processed by the node during the execution.
+ * @property eventId A string representing the event type.
+ */
+@Serializable
+public data class AIAgentNodeExecutionStartEvent(
+    val runId: String,
+    val nodeName: String,
+    val input: String,
+    override val eventId: String = AIAgentNodeExecutionStartEvent::class.simpleName!!,
+) : DefinedFeatureEvent()
+
+/**
+ * Represents an event indicating the completion of a node's execution within an AI agent.
+ *
+ * This event is triggered when a specific processing node, identified by its name,
+ * concludes its execution. It encapsulates details about the node's name,
+ * the input it operated on, and the output it produced.
+ *
+ * @property nodeName The name of the node that finished execution.
+ * @property input The input data provided to the node.
+ * @property output The output generated by the node.
+ * @property eventId A unique identifier for the event type.
+ */
+@Serializable
+public data class AIAgentNodeExecutionEndEvent(
+    val runId: String,
+    val nodeName: String,
+    val input: String,
+    val output: String,
+    override val eventId: String = AIAgentNodeExecutionEndEvent::class.simpleName!!,
+) : DefinedFeatureEvent()
+
+/**
+ * Represents an event that signifies the occurrence of an error during the execution of a specific node
+ * within an AI agent's process.
+ *
+ * This event captures information about the failed execution, including the identifier of the run,
+ * the name of the node where the error occurred, and the details of the error itself. It provides
+ * contextual information necessary for debugging, tracing, or analyzing issues encountered during
+ * the AI agent's execution.
+ *
+ * This event extends the `DefinedFeatureEvent` class, inheriting common event properties such as
+ * timestamps and message types.
+ *
+ * @property runId A unique identifier associated with the specific run of the AI agent.
+ * @property nodeName The name of the node where the error occurred.
+ * @property error An instance of `AIAgentError` containing the error details, such as a message, stack trace, and optional cause.
+ * @property eventId A string representing the event type.
+ */
+@Serializable
+public data class AIAgentNodeExecutionErrorEvent(
+    val runId: String,
+    val nodeName: String,
+    val error: AIAgentError,
+    override val eventId: String = AIAgentNodeExecutionErrorEvent::class.simpleName!!,
+) : DefinedFeatureEvent()
