+++ agents/agents-core/src/commonTest/kotlin/ai/koog/agents/core/feature/message/FeatureMessageProcessorTest.kt
@@ -2,6 +2,7 @@ package ai.koog.agents.core.feature.message
 
 import ai.koog.agents.core.feature.mock.TestFeatureEventMessage
 import ai.koog.agents.core.feature.mock.TestFeatureMessageProcessor
+import ai.koog.agents.core.feature.model.FeatureStringMessage
 import ai.koog.agents.utils.use
 import kotlinx.coroutines.test.runTest
 import kotlin.js.JsName
