+++ agents/agents-core/src/jvmTest/kotlin/ai/koog/agents/core/feature/writer/FeatureMessageFileWriterTest.kt
@@ -1,7 +1,7 @@
 package ai.koog.agents.core.feature.writer
 
 import ai.koog.agents.core.feature.message.FeatureMessage
-import ai.koog.agents.core.feature.message.FeatureStringMessage
+import ai.koog.agents.core.feature.model.FeatureStringMessage
 import ai.koog.agents.utils.use
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.joinAll
